Fix floating point exception on 64-bit Unix when using 6-Series airfoils.
The F2C converted 6-Series airfoil code returns its results to VSP
through an extern C struct.  The definition of this struct must exactly
match on the C++ VSP side.

On 32-bit Unix and 32/64-bit Windows, int and long are both 32-bit.
However, on 64-bit Unix, int is 32-bit and long is 64-bit.

The VSP side of the code uses ints, so the straightforward fix is to
change the typedef in the F2C code on the 6-Series side to match.