Scale triangles before running triangulate in CompGeom.
CompGeom is generally very robust, but when it does crash it is often
traced to the Delaunay triangulation in triangle.c.  These failures were
found to generally occur with sliver triangles of very small dimension.

This code scales the area to be triangulated to be on a scale of -.5 to .5
in each dimension.  This scaling improves the condition of the triangulation
problem and makes triangle.c more robust -- thereby making CompGeom more
robust.