Extrapolate Bezier surface when at or beyond maximum parameter.
Rather than force the value of the final control point, use the blend
functions to perform the Bezier surface calculations at or beyond the
maximum parameter value.

This change ensures the derivative calculations are correct up to the
edge of the surface.  The previous code would result in the control point
value for all derivatives at the surface limit.

The Bezier calculation should be reasonably well behaved for small
extrapolations.