Introduce Surf::TargetLen to calculate curvature based target mesh length.
Calculate the edge length which will result in a specified gap between
the edge midpoint and the surface.

This uses the larger magnitude principal curvature (k1) to ensure the
locally worst-case situation is handled.  The radius of curvature is
calculated r=1/k.  The local edge length is calculated which will give
the desired gap for the radius.

When the gap is larger than the radius, the target edge length is twice
the gap.

The flat surface case is handled through the large default value for len
and the check for small k1 -- if( fabs(k1) > tol ).

The curvature calculation returns NAN when both parameters are at their
maximum values.  i.e. in one corner of the surface.  This case is trapped
and the curvature values just inside the corner are calcualted instead.