Curvature meshing calculated on number of segments to represent a circle.
The number of segments required to represent a circle of a certain radius
determines the length of that segment.  In fact, for a given number of
segments, the segment length is a fixed fraction of the radius.

This code causes target edge length to be calculated as a desired fraction
of the local radius of curvature.  In this change, that fraction is hard
coded to 0.0981.  The formula for calculating this fraction is:

radfrac = 2.0 * sin(pi/N)

Where N is the number of segments desired to represent a circle. Using
N=64 results in radfrac=0.0981.