Add new global mesh control parameters to GridDensity object.
These parameters control the curvature based meshing.

m_MinLen -- Specifies the minimum edge length in the model.
m_NCircSeg -- Specifies the number of segments to represent a circle.
m_MaxGap -- Specifies the maximum gap between an edge and the surface.

m_RadFrac -- Used to implement the circle segment constraint.  This value
is calculated based on NCircSeg and should not be set directly.  Instead,
use SegNCircSeg to set this parameter.