Distribute points along a curve by integration.
The original curve point distribution algorithm can be viewed as an
approximation of this algorithm.  In this algorithm, the inverse of the
target distance (the point density) is integrated to distribute points.

This approach has the advantage (in effect) that the average target
length over a region is used to set the edge length for that region.
In the previous approach, the target length at the end of one edge was
used to set the next edge length.

The averaging effect of this approach should help to smooth out the
influence of fluctuations in the target length.

This approach also avoids the need for the U search process.  This
integration could be performed at the time the U,distance table is
constructed, but the target_len vector needs to be growth limited before
this can be done.