Add target edge length smoothing routines.
These routines limit the relative growth of edges to a fixed growth ratio.

The routine that takes a single edge as a parameter limits the size of that
edge compared to the surrounding edges.

The routine that takes no parameters first sorts all edges by size and
then walks the list from smallest to largest.  Each edge neighbor is
compared to make sure the growth isn't too fast.