Introduce simplified BorderTesselate()
The borders are tesselated to create ISegChain versions of the
borders - which are later converted into SCurves - which are
tesselated for use in the mesh.

Consequently, the tesselation in LoadBorderCurves needs to
represent the curve well, but it does not directly lead to the
points in the final mesh.

In BorderTesselate, the points are distributed uniformly in the
U,W space of the surface rather than according to the distribution
for the final mesh.