Introduce edge source builder and applyer.
These edge sources are used to communicate density information
from one intersection curve to another.  This is needed because
the curvature based meshing does not create traditional sources
and because the growth limiter needs to propagate this information
to other areas.

Other applications of the growth limiter propagate the information
through the mesh connectivity, but by the time that can be done,
the edges are already tesselated.