Back off on curve subdivision size.
The handbook of grid generation suggest 5-10x the eventual discretization
of the curve.  1/5 of 1000 still allows for 200 segments for a curve, which
seems reasonable.

The new edge sources introduce a computationally expensive step.  This
change reduces that cost by a factor of 10.  If a more elegant algorithm
is implemented instead, this change can be reconsidered.