Introduce edge source pruner.
Some of the edge sources will have no effect on the mesh because other
nearby edge sources produce smaller cells.  The pruner goes through all
the edge sources and removes those that can't possibly have an effect
on the mesh.

This optimization saves a second or two.  Not a lot, but it is worth
keeping now that it is written.