Check for degenerate parametric coordinates in CompCurvature
When the Bezier surface has one edge collapsed, that edge is degenerate
and the gradient along that edge will be zero.

This code detects that situation and calculates the curvature along the
non-degenerate curve instead of the true surface curvature.

In some situations, both U & W directions are degenerate.  Why this occurs
is not clear, but in that situation, the curvature is set to zero.