Improve handling of not-quite perfect curve tesselation.
Spread the error of the final tesselation point over the entire curve.
This is not as good as an elliptical solver, but it is very simple and only
requires a single pass over the data.

Although this should strictly be done in terms of the curve length, here
it is done in the u-parameter.  Because the adjustment is very small, the
differences are expected to be small.