Maintain accurate tmax and counter to trigger recalculation.
Ensuring that tmax matches the actual maximum source strength will
minimize the search radius required to check if this source is dominated.
Checking and re-setting tmax every time was very wasteful in some cases.

This code maintains tmax and a counter nmax with the number of occurances
of tmax.  When an instance of tmax is reduced, the counter is reduced.
When the counter hits zero, a new tmax is searched for and counted.