Change zero length edge heuristic.
This heuristic attempts to identify small edges which should not be used
to propagate mesh limiting.

Before, 0.7 was considered a reasonable value because t was assumed to be
fairly small.  A case was brought up with a mostly uniform fairly coarse
mesh where t was large.  Consequently, a small edge was not propagated when
it probably should have been.