Remove edge adjustment from edge sources.
The edge adjustment was an attempt to set an edge source strength (at the
center of an edge) which would yield the edge length when its influence
grew to the endpoints of the edge.

This calculation goes singular for GrowRatio >= 3.0.  In the end, this is
a bit of un-needed cleverness.