Fix for mesh map interpolation when generating half-meshes.
The intersection curves interpolate their mesh density from the underlying surfaces.

When we do a half model, a fake surface is created as the cutting surface.
That surface is never added to the 'master' set of surfaces -- so, it
never has a mesh density map made.

Two solutions were considered:

1) Make a mesh density map for the cutting surfaces.  Would happen in
CfdMeshMgr::IntersectYSlicePlane().

2) Detect the intersection surface (CompID = -999) and don't use it for
calculating the curve density.  This is the solution used in this patch.