Write out surface quadraliterals instead of triangles.
This should be slightly faster because no triangles are constructed and
the triangle nodes are not merged.

The quadrilateral nodes are automatically merged according to the topology
of the surface - edges do not wrap back on one another.  This means that
wing trailing edges now render correctly even though X3DOM does not obey
the creaseAngle parameter.