Change ptrID to be based on rand() instead of memory address.
ptrID needs to be changed to be constant and hopefully unique over the
life of an object -- to stay the same over the course of file load/save
and other operations.  Setting ptrID based on the memory address of the
object runs the risk of collisions.

This change converts initialization of ptrID to be based on a random
integer.  The random number generator is seeded in main() based on time()
which returns the number of seconds since 1/1/1970.