Remove unused Geom::id_str
Geom::id_str appears to go back to the RAM file format days and has been
replaced with ptrID.  However, it was still read/written and carried
around.

In addition, the base constructor sets the value based on a timestamp.
Consequently, opening and saving an identical model would result in a
changed file.  Removing this allows files to stay the same.