Introduce resetClipBoardPtrID() for Insert and Paste.
This zeros out the ptrID's of all components on the clipboard.  A zero
ptrID forces a new ID to be created for the component when needed.  This
will occur later, when something is written to file.  All actual
relationships between components are handled through actual pointers.

This causes an explicit paste or insert operation to create new unique ID's
for all the components involved.  This is strictly overly conservative, it
is only required when an ID collision exists.