Increase precision in XSec files
Some users were running into problems with small fuselage cross sections
which were traced to the limited precision output in XSec files.  This
change increases the field width to give all available precision.

This also changes to %e from %f format.  This forces scientific notation
which will ensure the digits presented are the most significant digits.
It is possible that this change will cause other programs that use this
file format problems.  Both the width of the fields is changing and the
use of scientific notation.