Don't special-case rounded wingtip leading/trailing edges.
The removed code forced the leading and trailing edge points of a rounded
wingtip to the exact expected solution.  However, when slats/flaps were
applied -- or if an airfoil file with shifted le/te was used -- this
would not produce appropriate results.

Removing this code lets the general case code handle these points just like
all the rest of the airfoil points.

This fix should close Issue #11:
https://github.com/OpenVSP/OpenVSP/issues/11