Handle XDDM Constants in addition to Variables.
The Cart3D optimization framework changes the type of some quantities
from 'Variable' to 'Constant' depending on the context.  The geometry
modeler needs to apply the values of either type no matter what.

Previously, VSP ignored anything that was not a 'Variable'.  This change
allows VSP to read/write/manipulate lists of Variables and Constants.