Design files are initiating sources of parameter setting, not from link.
The set_from_link and set variants differ in how they set a flag to enable
detection of circular sets of parameters.  Plain set() sets the flag to
indicate that it is the starting point.  _from_link sets the flag to
indicate that it is somewhere in the middle of a sequence.

The DES and XDDM files should have used the plain set() from the beginning.
I did not fully understand the differences when this code was written.