Make rigorous growth limiting optional
Add file and GUI options to control rigorous 3D mesh growth limiting.  This limiting
step is computationally expensive, but may be required for certain applications and
for certain geometries.

Geometries with facing (but nonintersecting) surfaces such as nacelles or high-lift
systems present this situation.

Applications which will need to grow a volume mesh in those facing gaps will need the
mesh to be limited across the gap.  Other applications may not have this requirement.