Fix MS_Wing chord update on scale.
The previous code assumed that the tip chord of one segment would update
the root chord of the next segment in order to maintain consistency.  This
assumption proved to be false.  Rather than trigger matchWingSects which
depends on the current active section, this fix simply sets both the
root and tip chords.

Bug reported by Bill Fredericks.