Update example files to latest VSP file format.
The example files had become far out of date as new pieces of information
were added to the VSP file format.  On load, VSP would throw a warning for
missing information and fill in a suitable default.  This causes a large
number of messages to display to the console -- which is unnerving to new
users.

This change was performed by simply opening the old files and saving them
with a current version of VSP.