Fix CG contribution of point mass components.
Despite its name, getTotalTranVec() only calculates the translation of the
component within its own coordinate system.  If the component is connected
to another component, this does not result in the coordinates in absolute
space.

Instead, establish a point at the origin of the local system (a default
vec3d) and then transform it to global coordinates.  The transformation
matrix used already incorporates getTotalTranVec().