Add ISegChain::Match to test for matching chains.
Match checks if two chains are coincident.  To be coincident, the chains
must use the same A surface.

In addition, the first and last points in the chain must match within a
small tolerance.  Not all points in the chain are checked -- only the
first and last.  Flipped chains are checked as well.

The CompPnt() routine averages the coordinates of the point based on the
various surfaces which underly it.  This may cause robustness problems and
should perhaps be changed to use only the A surfaces which have already
been matched.  At that point, comparison may be made on u,w coordinates
instead of x,y,z.