Introduce split sources to help mesh quality near intersections.
The finite surface background mesh sometimes is insufficient near
intersections.  This is not a significant problem for the surface field
mesh because the growth ratio is enforced through connectivity.  However,
the curves are tesselated based entirely on the distance table.

Split sources are created at the ISegChain splits -- points where a
chain was split.  These correspond to intersection curve endpoints that
do not correspond to surface boundaries.

Split sources are initialized by one of the local surfaces and are updated
by the distance table endpoints -- or they can update the distance table
endpoints.