Keep track of non watertight entities in CFD Mesh
This change stores all the nodes from the watertight check so they will
still be available for visualization of bad edges/tris.  The memory impact
is not great because these nodes are freed before the next mesh is generated.
I.e. they only exist when you aren't generating a mesh.  That said, it
does imply a duplicate set of nodes to those used for visualization.

A more efficient approach would be to add a 'bad' flag to the primary
edge/tri storage and modify the visualization to use that flag.