Bump version to 2.3.0
Version 2.3.0 contains one significant feature and a small bug fix.
Version 2.3.0 is likely the last 2.X feature release for VSP.  We will
continue making 2.3.Y bug fix releases for a long time.

For some time, almost all developer effort has been focused on the pending
3.X project.  VSP v3 is a near total rewrite keeping everything you love
about VSP, but ushering in new features galore and a more maintainable
code base.  We hope to release a not-for-production-use alpha version of
v3 for users to begin testing soon.

Fixes/features in this release:
 - Add ability to model outer domain boundary in CFD Mesh.
 - Match multi-body edge shares in CFD Mesh.  (Thanks Sriram R.)

You can now generate a CFD Mesh including the outer boundary.  Some volume
grid generators need to start with a fully enclosed mesh, so we now support
that mode of operation.  The outer boundary can be either an explicit VSP
component or a box generated for you.  You can model the full domain, or
just the +Y half domain -- VSP will mesh the symmetry plane for you.  There
are separate mesh size controls which apply to the outer boundary.

In CompGeom, VSP tries to match pairs of non watertight geometries to form
a single watergith geometry.  In CFD Mesh, VSP tries to pair and stitch
any perfectly coincident boundary edges.  In cases where there were three
or more components that daisy chained together, VSP do this incorrectly.
With this bug fix, you can now daisy chain an arbitrary number of
components to form a watertight body -- so long as the boundaries match
exactly.