OpenVSP 2.9.0, first public alpha for version 3.
For the past two years, a dedicated team has been re-writing VSP from
the ground up.  This is the first public release from that effort.

The core v3 team responsible for this work so far is:
Alex Gary, J.R. Gloudemans, Yue Han, David Marshall, and Rob McDonald

The 2.9 series will be the alpha and beta versions leading to 3.0.0, which
will be the first production release for this branch.

Unfortunately, this development was done in a way that the development
history is not a meaningful continuation of the 2.X history.  Consequently,
this commit starts a new history from the beginning.  Do not try to work
across the branches, no good can result.