Add -fPIC flag for x86_64 on Linux.
The Position Independent Code flag is needed for shared libraries on some
platforms.  This is normally implied for shared libraries, but we build
most of VSP as static libraries, so it is not set automatically.

Later, we use these static libraries to build a shared library for the
Python bindings.  So, for dual-use libraries, we need to set -fPIC manually.