OpenVSP 2.9.1, second public alpha for version 3.
Long overdue, this second alpha has tons of changes from 2.9.0.  This
version still has lots of bugs and missing features, but it should be a
dramatic improvement -- or at least a much better preview of 3.0.0.

Major improvements include:
 - Subsurfaces
 - Custom Components
 - Greatly expanded API
 - Preliminary Wing and Stack Geom types
 - Airfoils and other XSec types implemented
 - Pick Geom (p) and Pick rotation (r) modes
 - Rulers, materials, and lighting
 - Lots of bug fixes

The 2.9 series will be the alpha and beta versions leading to 3.0.0, which
will be the first production release for this branch.