Fix DegenGeom problem with wing parameterization reversal in v3 vs. v2.
In v2, the around-the-airfoil parameter started at the TE, and
wrapped around the LE back to the TE -- but it traversed the
top surface first.

In v3, the parameter traverses the bottom surface first.

DegenGeom cares in places where it uses top/bottom concepts
distinctly.  This swaps each of those places.