OpenVSP 2.9.2, third public alpha for version 3
Long in coming, but worth the wait.  This third alpha has many and vast
improvements from 2.9.1.  While there are still bugs and small feature
holes, all the major features are in there.  This is getting close to what
3.0.0 will be.

Major improvements include:
 - Degenerate geometry support to enable low-order analysis tools.
 - Fuselage and stack component skinning
 - View clipping
 - Subsurface visualization and polishing
 - Active XSec highlighting & view
 - Saving CFDMesh settings
 - Custom geometry improvements
 - Start of advanced parameter linking
 - Fix wing sweep/twist interaction, fix dihedral implementation
 - Fix problems with JPEG files

The 2.9 series will be the alpha and beta versions leading to 3.0.0, which
will be the first production release for this branch.