Introduce SimpleSources to actually interact with CFDMesh.
Originally, Sources were copied to handle symmetry every time UpdateSources
was called.  Since sources contained numerous Parms, this seemed to wreak
havoc.

So, SimpleSources are now the destination when Sources are copied for
symmetry.  The parameters are placed in plain doubles, with no registration,
ID, etc.  SimpleSources are used by CFDMesh and for visualization.  The
original source class is used for interacting with the GUI, file I/O, etc.