Change wing capping parameter treatment to allow sources and subsurfaces on caps.
CompGeom and CFDMesh have been modified to remove coincident wing root
cap surfaces after u,w parameters have been assigned for later use by
subsurfaces.  Consequently, no unexpected parameter jump occurs for those
cases.

Chord values are clamped from never going to zero (removed in subsequent
change).  When that is removed, capping can fail when chords go to zero.
This is not a problem for the tip cap case (the common case) because losing
the highest extent of the parameter space is not a problem.

So, the only situation that causes unexpected jumps in parameter space is
when the root chord of the wing goes to zero when a root end cap was turned
on.