Use XSecCurve ComputeArea for DragFactors rather than XSec.
Although this algorithm still works with 3D points, it could instead be
done with 2D points from XSecCurve.

This corrects a problem for out-of-plane center points in the curve area
calculation.  Could also be a problem if the transformation took the area
so the origin was not contained by the area.

A 2D algorithm could handle more complex shapes by using a signed area
calculation.