OpenVSP Version 3.1.0
OpenVSP 3.1.0 brings the much awaited release of VSPAERO, the new vortex
lattice solver that has been developed to be a part of OpenVSP.  There
are a few other improvements and fixes, but I'm certain everyone will
want to update to get their hands on VSPAERO.

VSPAERO is a thin-surface code for inviscid subsonic and supersonic
aerodynamics.  It is multithreaded to make the most of common multi-core
computers and uses a mesh agglomeration scheme to improve scaling with
large meshes.  The matrix is solved with an iterative GMRES solver.
VSPAERO includes a simple actuator disk model to represent propulsion-
airframe interaction and also the ability to calculate common stability
derivatives.

Improvements and additions since 3.0.4
 - VSPAERO vortex lattice solver and viewer
 - VSPAERO integration into OpenVSP GUI
 - Initial use of Travis-CI continuious integration build/test server
 - Search multiple paths for CustomScripts, working, home, then vsp's location
 - Add home and vsp's location to favorites pull-down in file select dialog

Bug fixes
 - Reading wings with sweep location=1.0 from file (reported by Nick Borer)
 - Wing parameter limits now allow tiny segments (reported by Nick Borer)
 - Fix Linux build problem introduced at the last moment before 3.0.4