OpenVSP Version 3.2.0
OpenVSP 3.2.0 is the first feature release in a while -- there is a lot
of great stuff included to make it well worth the wait.

Blunt trailing edges -- Wing section TE's can now be blunt.  Airfoil files
with a gap between the TE points will be left that way (re-read your files
on existing models to get this behavior).  Airfoil curves can be modified
to thicken or sharpen the TE; they can also be trimmed to blunt the TE.
The sharp corners have parameter values W=0.001 and 0.999, use these values
when creating subsurfaces or attaching to the corner.

A negative volume flag has been added to components to cause them to poke
holes in CFDMesh.  In all other cases, these components have normal
behavior.  This should be useful for modeling engine flow-through.

Actuator disk support has been added.  The disk is described in DegenGeom
for VSPAERO and a thin membrane is meshed in CFDMesh.  Some volume meshers
should be able to mesh up to both sides of the disk.  The actuator disk
is immune to negative volume, so you can put a disk in your flow-through
duct if you want to.

VSPAERO has some great improvements as well.  Easy actuator disk
integration, velocity surveys, and some load integration bug fixes.

Improvements and additions since 3.1.2:
 - Blunt TE airfoil support
 - Negative volumes for CFDMesh
 - VSPAERO auto-setup of actuator disks
 - VSPAERO field velocity survey calcualtions
 - Actuator disk component (example custom component)
 - Improved actuator disk integration with VSPAERO
 - Control surface subsurfaces
 - Drag-N-Drop for picking parameters
 - Save/Import for FitModel optimization problem formulation
 - Programmatic GUI for CFDMesh, Parameter Linking, CompGeom, and DegenGeom
 - Custom component support for actuator disk and negative volumes

Bug Fixes:
 - VSPAERO Fix some pressure integration bugs
 - Split symm plane to improve CFDMesh mesh quality with huge domains
 - Tweak tolerance for short intersection edges in CFDMesh
 - Set z-buffer distance dynamically to handle large models

One last note, with this version we also add two new developers to the
team, thanks much to Brandon and Mitch for their contributions.