Move CompGeom orientation consistency check until after reverse transform.
Delaunay triangulation is performed in YZ plane.  Triangles are transformed
into that plane, triangulated, and then transformed back.

The newly created triangles are not created in a consistent orientation, so
they are compared to the original triangle's orientation and changed as
needed.

Checking the orientation with a dot product in the original coordinates is
more robust than the old test in the transformed coordinates.

Thanks to Erik Olson for the test case to chase this down.