OpenVSP Version 3.3.1
Small but mighty!

OpenVSP 3.3.1 may 'only' be a bugfix release, but it fixes some really
big and longstanding bugs.  The fixes are so significant that I
considered calling it 3.4.0 -- or even 4.0.0!

Whatever you do, update to this version, you'll be glad you did.  Shaded
rendering is significantly improved, a few performance improvements, and
the problems with 6-series and file airfoils have been fixed.

Bug fixes:
 - Fix various problems with 6-Series and File Airfoil types.
 - Multiple normal vectors at corners and edges for better shading.
 - Improve feature line rendering to reduce artifacts.
 - Eliminate transparent surface rendering artifacts.
 - Tweak adaptive feature lines to improve consistency of results.
 - Add OpenVSP version number to output IGES files.
 - Tweak target mesh length calculation to speed CFDMesh.