OpenVSP 3.5.1
This version has lots of fixes and a handful of features.  There is
something here for everyone, so update -- you don't want to be left out.

The addition of tessellation clustering control for wings (chordwise and
spanwise) will be important to lots of users.  This only affects the
wireframe and things related to it (DegenGeom, VSPAERO, CompGeom, etc).
It does not affect the underlying shape or the surface parameterization.

To do this, NACA 4-digit and CST airfoils types had to be re-parameterized
to an equal-arc-length basis.  This will change loft with mixed airfoil
types and possibly attachment points.  An option to recover the old
behavior is included.

Cluster is controlled with two parameters (LE/TE or Root/Tip).  If you
set them =1.0, you get uniform spacing (the default).  Smaller numbers
cluster, larger numbers spread things out.  Feedback on the level of
control and the growth rate is appreciated.  This may need tweaking down
the road.

The libJPEG library has given us no end of trouble over the years and I've
finally had enough.  I ripped it out and have replaced it with a smaller
and simpler library.  Hopefully this fixes all the problems users have
had with image files.  This will be mostly user invisible, but you can
no longer do JPEG screenshots -- PNG only.  Also, you now have your choice
of image formats for background images (JPG, PNG, GIF, BMP, TGA).

Features
 - Chordwise tessellation clustering control
 - Spanwise tessellation clustering control
 - Equal arc length skinning for four digit and CST airfoils
 - Add selection Show Only button.  Suggested by Jason W.
 - More file formats supported for background (jpg, png, bmp, gif, tga)
 - Make XSec highlight curves adaptive
 - Improve polygon offset to improve feature line rendering

Fixes
 - Fix update custom components change position/rotation. Thanks Erik O.
 - Fix problem with wing list update in VSPAERO GUI
 - Fix recent problems with logslider
 - Fix fuselages with many file XSecs
 - Fix O(n^2) problems in fuselage import and read
 - Get rid of libJPEG