OpenVSP Version 3.5.2
Another collection of small fixes and features, with one or two that are
sure to be crowd pleasers.  Get it now and don't be left out!

Most of the changes fall into two categories.  Sub-Surfaces and the main
GUI.

The first sub-surface work allows multi-surface geoms (common with custom
components) to assign sub-surfaces on a per-surface basis (rather than to
all surfaces).

In addition, sub-surfaces are now output to DegenGeom files.  This kind
of extension of the DG file format was always expected, so hopefully
parsers have been written to 'just work'.

Sub-surfaces are output as a u,v polyline, it is up to the DegenGeom user
to figure out what that means for their stick or plate representation.
Sub-surface names are also written out as well as a flag indicating which
side of the polyline is the 'inside' of the subsurface.

The main window has been refactored such that it is now generated on the
fly.  This will make future updates much easier.  While I was there, I
made a bunch of other improvements and cleanups.

The Undo button is now located under the Edit menu, and Ctrl-Z
(Cmd-Z on Mac) will trigger Undo.  Lots of menu commands now have sensible
shortcut keys that make common actions easier and also document some
little-known capability.

On Mac, the main menu bar now appears in the top window bar like other
native applications.  This may take some getting used to.

If you're still looking for a reason to update, here it is.  With this
version, a press of the 'f' key will fit the visible model (zoom and
center) to the screen (and change the rotation center to the center of
the visible model).

Features
 - Per-surface sub-surfaces
 - Sub-surfaces written to DegenGeom files
 - Programmatic main window
 - Fit visible model to window with 'f' key
 - Menu rearrangement with lots of shortcut keys

Bug fixes
 - Fix CFDMesh problem with vertical tails (thanks Carmine Vassallo)
 - Fix problem passing shortcut keys to all listeners