OpenVSP Version 3.6.1
One big feature, a handful of enhancements, and a bunch of bug fixes make
this version well worth the update.

The big feature is a projected area calculator.  Use it to calculate
the projected area of a component or a set of components.  Project in the
X, Y, or Z directions -- or along an arbitrary vector -- or along a
vector normal to the u=0, w=0 point on a given component.  You can also
bound the projected area by a component or a set.

The resultant area is triangulated and added to the model as a MeshGeom for
visualization or export to STL.  Although only the final area is reported
in the GUI, component areas and the area outline are available in the
output file or via the API.

So, whether you want frontal area for a drag estimate, rotor downwash area
for a download calculation, or you just want the projected area of a
complex wing made of multiple components -- you're in business.

The super ellipse and rounded rectangle cross section types have been
enhanced to creation of a bunch more related shapes.  Thanks to Frank
Zhang for those enhancements.

The bug fixes include some old and some new.  All good stuff.

Features:
 - Projected area calculator with many use cases
 - Enhanced superellipse curve type - Frank Zhang
 - Enhanced rounded rectangle curve type - Frank Zhang

Bug Fixes:
 - Protect against NULL pointer dereference from API or Custom Components
 - Fix crash with airfoils trimmed at barely different spots (thx Jason W.)
 - Fix longstanding bug with handling of rounded corners for General XSec
 - Fix deactivated sliders not updating
 - Set unused capping parameters to value of least surprise