OpenVSP Version 3.7.0
This version includes two significant new features that have been in the
works for some time.  These are the estimation of supersonic wave drag
by the area rule and a capability to save sets of variable settings in a
given model.

The wave drag capability has been much requested and long awaited.  Since
its demonstration at the 2015 Workshop, a number of slick improvements have
been made.  Thanks to Michael Waddington for all the work and to Sriram
Rallabhandi for the C implementation of the Eminton Lord calculations.

The variable preset capability has a number of use cases; it allows you to
save the model state and return to that state quickly and easily.  Perhaps
you have a model with takeoff, landing, and cruise configurations.  Perhaps
you adjust the tessellation for VSPAERO, but want different resolution for
mass properties, and different again for Cart3D.  Or, if you're doing a
design study, you may want to save the baseline model and different
optimal solutions (weight optimal, fuel optimal, cost optimal, etc).
Thanks to Bryan Schmidt for all the work on this feature.

Features:
 - Wave drag estimation by area distribution and Eminton Lord calculation
 - Variable presets for saving model state