OpenVSP Version 3.8.0
Don't forget to register for the OpenVSP Workshop (Aug 23-25)!

This version is more focused than many of our releases.  It is really
all about upgrading the VSPAERO integration (and one bugfix).

The VSPAERO GUI has been comprehensively updated.  You can now run alpha/
beta/Mach sweeps from the GUI.  Results are automatically parsed and are
ready for review from the GUI.  VSPAERO is now available through the
Analysis Manager in the API.  The results are parsed and are available
through the Results Manager in the API.

In addition, a bug fix for files saved since version 3.7 that use super
ellipse cross sections.

Features:
 - Overhaul of VSPAERO integration
 - VSPAERO alpha/beta/Mach sweeps from GUI
 - Visualize VSPAERO results in GUI
 - VSPAERO available through API
 - Changed default wing LE/TE clustering to 0.25
 - Build on OSX on Travis-CI
 - Improve automated testing

Bug fixes:
 - Duplicate parameter name in superellipse cross section