Update Matrix4d::rotate to apply on-top of current transformation.
This makes rotate behave like rotateX,Y,Z and other transformations.

However, the only in-tree caller that depended on the existing behavior
was the skinning-related transformations in XSec.cpp.

Note, this method is exposed to the API, so there may be custom component
or scripts whose behavior will be changed by this.