OpenVSP Version 3.8.1
Propellers for all!

This release adds a new propeller component and fixes a pair of unrelated
bugs.  Though it may seem like a small change, the propeller component
constructs its surface differently from other components.  It represents
a surprisingly large amount of work.

Instead of defining the blade shape (chord, twist) at specified stations,
these are instead specified as continuious curves.

The propeller component comes with a bunch of related accessories -- the
blade activity factor is calculated on-the-fly, there is a one-click option
to switch to a counter-rotating prop, propeller folding is built-in -- with
a user-defined fold axis, there is a blade element export/import
capability, and the BEM data is available through the API.

Features:
 - New propeller component
 - Blade element import/export and API

Bug Fixes:
 - Fix DegenGeom crash on Matlab export
 - Fix GUI output of mass properties calculation