Use geom type lookup based on names to prevent problems with duplicate nums
Built-in Geoms have two names that must be available for lookup.  The mixed
case name is generally assigned in the Geom's constructor and identifies
the Geom type in the XML file.  The ALL-CAPS name is generally assigned in
Vehicle::Init() in m_GeomTypeVec and is used to identify the type from
the GUI and the API.