Disable compilation of Cartesian example.
For some reason, this causes the build to fain on Linux due to
missing -lpng -lz -lXft -lfontconfig -lXinerama that it does
not complain about when the full vsp executable is built.

This could also be satisfied by supplying these libraries, but
this workaround is preferred to keep dependencies to a minimum.

This did not used to occur, it is unclear what change caused this
behavior.