Adds Mass Properties Test Script with 7 Geom Tests
This test script is similar to the APITestSuiteMassProp. The script
contains 7 functions that differ mainly by the type of geometry
created. Each test begins by generating a basic geometry that has well
established analytical mass properties. The MassProp is then executed
and the results are displayed, along with the analytical results. These
tests also serve to demonstrate how to create various geometries and
run analyses using a script.

*Note: The MassProp Analysis results will be incorrect for shell
geoemtries at this point.